backend/pyproject.toml
[tool.poetry]
name = "athena-sentinel-backend"
version = "0.1.0"
description = "Athena Sentinel backend"
authors = ["you <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
SQLAlchemy = "^2.0.13"
asyncpg = "^0.27.0"
python-dotenv = "^1.0.0"
passlib = "^1.7.5"
python-jose = "^3.3.0"
slowapi = "^0.1.4"
aiohttp = "^3.9"
feedparser = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
httpx = "^0.24.0"
